Simpler defaults for InteractionCaptureLayer.
On OS X, shift + scroll with a scroll wheel is a systemwide shortcut for horizontal
scroll, which means that there's no `deltaY` to inspect on the event. Instead of
trying to be clever here, I'm just going to make the defaults "just work" and leave
this commit message to document the behavior.