Fixes to the AnimateProps mixin/decorator.
It now animates from the previous animated value instead of the previous prop. This is so
multiple prop updates in quick succession (say, if the render is sloppy and renders multiple
times with value-equals and reference-unequals objects) will still animate smoothly from
wherever the view currently is.

Defensive use of _.cloneDeep, because d3 explicitly does not defensively clone and this has
the potential to cause seriously bad side effects.