Fix bug where resetting `ChartProvider`'s `seriesIds` could overflow the stack.
Add tests for the reducer to catch this case in the future.

The bug would happen if you did three things;

- Set the `seriesIds` prop to an inline array literal, thereby bypassing reference-equality checks.
- Provided any of the callback methods, such as `onYDomainsChange`.
- In the callback method, you did something that caused a rerender, such as setting state.

The bug happens like so:

1. On initial render, initialize the store and kick off any callbacks.
2. Callback sets state or otherwise causes a rerender.
3. render passes array literal to `ChartProvider`, which dispatches `setSeriesIds`.
4. Reducer recomputes a bunch of state for the new IDs, but the only thing that changes
   is the outermost object.
5. Callbacks fire because the state changed.
6. Go to 2.

The simple solution is to ignore any request to change `seriesIds` in a deeply-equal way.

Additionally, a related by marginally less severe issue caused by reference-unequal
values for `loadData` can only realistically be avoided by warning people that they
MUST provide reference-equal values for it if they don't intend to change it. The docs
have been updated to mention this.

This bug is truly caused by a deeper, two-way-binding issue. `ChartProvider` is mediating
reads and writes in two directions using two different patterns, and is rather
error-prone. I tried to repair the underlying issue, but it's nontrivial. This solution
seems reasonable, because setting `seriesIds` is currently the only way to actually get
into this loop. Making the loop impossible in the first place is tricky. The reason that
this is the only setting that can cause the loop is as follows:

- This can only be caused by props that interact with the store, either be reading from
  it (in the case of "on change" functions) or writing to it (in the case of "controlled
  props").
- Six props can be pushed to the store, and six values can be read from the store into
  a callback.
- Of the six values read from the store, four are NON-overridden UI state, and the other
  two are load state/results.
- Of the six props that are written to the store four are overridden UI state, one
  triggers a deferred load (`loadData`) and and the last is `seriesIds`.
- The four reads and four writes that are UI-related don't interact with each other, since
  they deal with different underlying settings (override versus not).
- The two reads and one write that are loading-related can indeed cause a loop, but
  because it's necessarily debounced/deferred and because we can't actually examine the
  culprit (i.e. `loadData`) for deep equality, the best we can do is documentation.
- `seriesIds` is the last remaining option: before this change, a deeply-equal value
  used to cause a rewrite of the outermost objects for (among other things) the NON-
  overridden UI state. It wouldn't change the object's values in this case, however.
- That overwrite would trigger the callbacks (unlike the UI-related props) and those
  callbacks had the potential to synchronously loop (unlike the loading-related props).

Therefore, the only combination of props that could cause this issue are `seriesIds` and
any of the UI callbacks. `loadData` and friends can cause a slow-motion infinite loop
that is deferred enough to prevent crushing disaster like this stack overflow, but still
sucks (and is therefore documented).

This fix is unfortunately a bit fragile in that it doesn't fix the underlying issue,
and any new props or interactions between these props might reintroduce it. Hopefully
this commit message and the slightly unusual deep-equality check in the reducer are
warning enough.