Clean up exported typings. Fixes #15.
- Replace { min, max } with Interval.
- Adjust casts on propTypes and defaultProps so that they
  1. don't cause Props to be inferred as `any` in some scenarios (i.e., wrapWithAnimatedYDomain).
  2. don't leak anything besides being a React.ValidationMap<Props> or Props, as appropriate.
- Cast selectors that use lodash so the returned type is not a _.Dictionary, but a keyed object.
- Make a bunch of private methods actually private.
- Type some functions more specifically or remove any-casts to prevent any-propagation.

Some color on the defaultProps and propTypes casts. For some silly reason, the typings for
defaultProps and propTypes are both totally useless. In the case of defaultProps, it's typed
as P (that is, props from the enclosing class), which is useless in practical terms, since
you generally provide defaults for the props that are optional, meaning that you're defining
the exact inverse of P for the value. In the case of propTypes, it wants something with an
index signature, but literals can't have index signatures (AFAICT), so you always have to
cast it to another index-signature-able type to make it check out. Furthermore, if you cast
either of these to `any` and then run them through a function like `wrapWithAnimatedYDomain`,
Typescript seems to propagate that right through to the Component's type itself, so even
though you explicitly type it as `React.Component<Props, State>`, it ends up as
`React.Component<any, State>`, which doesn't help anyone.